+++ libraries/RS485/RS485.cpp
@@ -2,7 +2,7 @@
 //    FILE: RS485.cpp
 //  AUTHOR: Rob Tillaart
 //    DATE: 30-okt-2017
-// VERSION: 0.3.0
+// VERSION: 0.4.0
 // PURPOSE: Arduino library for RS485 modules (MAX485)
 //     URL: https://github.com/RobTillaart/RS485
 
@@ -20,7 +20,8 @@ RS485::RS485(Stream * stream, uint8_t sendPin, uint8_t deviceID)
   _deviceID = deviceID;
 
   pinMode(_sendPin, OUTPUT);
-  setRXmode();    //  receiver mode ==>  _sendPin == LOW
+  //  default receiver mode ==>  _sendPin == LOW
+  setRXmode();
 }
 
 
@@ -87,12 +88,14 @@ size_t RS485::write(char * array, uint8_t length)
 //  smallest and slightly fastest.
 size_t RS485::write(uint8_t * array, uint8_t length)
 {
-  uint8_t n = 0;
+  setTXmode();       //  transmit mode
+  size_t n = 0;
   for (uint8_t i = 0; i < length; i++)
   {
     n += _stream->write(array[i]);
     yield();
   }
+  setRXmode();       //  receiver mode
   return n;
 }
 
@@ -146,25 +149,29 @@ size_t RS485::write(uint8_t * array, uint8_t length)
 //     EOT          end of transmission
 //
 
-void RS485::send(uint8_t receiverID, uint8_t msg[], uint8_t len)
+size_t RS485::send(uint8_t receiverID, uint8_t msg[], uint8_t len)
 {
+  size_t n = 0;
   uint8_t CHKSUM = 0;
+
   setTXmode();                 //  transmit mode
   _stream->write(SOH);
-  _stream->write(receiverID);  //  TO
-  _stream->write(_deviceID);   //  FROM
-  _stream->write(len);         //  LENGTH BODY
-  _stream->write(STX);
+  n += _stream->write(receiverID);  //  TO
+  n += _stream->write(_deviceID);   //  FROM
+  n += _stream->write(len);         //  LENGTH BODY
+  n += _stream->write(STX);
   for (int i = 0; i < len; i++)
   {
-    _stream->write(msg[i]);
+    n += _stream->write(msg[i]);
     CHKSUM ^= msg[i];          //  Simple XOR checksum.
   }
-  _stream->write(CHKSUM);
-  _stream->write(ETX);
-  _stream->write(EOT);
+  n += _stream->write(CHKSUM);
+  n += _stream->write(ETX);
+  n += _stream->write(EOT);
   _stream->flush();
   setRXmode();                 //  receive mode
+
+  return n;
 }
 
 
@@ -275,6 +282,17 @@ bool RS485::receive(uint8_t &senderID, uint8_t msg[], uint8_t &msglen)
   return false;
 }
 
+
+size_t RS485::send(uint8_t receiverID, char msg[], uint8_t len)
+{
+  return send(receiverID, (uint8_t *)msg, len);
+}
+
+
+bool RS485::receive(uint8_t &senderID, char msg[], uint8_t &len)
+{
+  return receive(senderID, (uint8_t*) msg, len);
+}
 
 
 //////////////////////////////////////////////////